Unofficial multisig-tool fork to extract Bcash stuck in your Coinbase vault.
In case you believe:
1) Coinbase will add support for extracting your unspent Bcash from your vault, and
2) Bcash is still going to be worth something significant when that happens
Then I recommend you wait for the official Coinbase support rather than use this branch.

I've personally succesfully used this tool to extract my coins.
HOWEVER, I do not ensure the tool works properly or that you would not lose your coins as a result of using it. Use at your own risk and liability.

If you've decided to use the tool, please be sure to:
- Correctly fill in your unspent output address in insight.js.
- Fill in the correct information in address_data.json (in the same format returned by insight.bitpay.com/api).
- Triple-check the resulting transaction before broadcasting it to the Bcash network.